fix: keep account schema client-safe
