+++ demos/data-model-forms/app/assets.ts
@@ -1,17 +1,18 @@
 import * as path from 'node:path'
 import { createAssetServer } from 'remix/assets'
 
-import { assetsBase } from '../routes.ts'
+import { assetsBase } from './routes.ts'
 
 const isDevelopment = process.env.NODE_ENV === 'development'
 
 export const assetServer = createAssetServer({
   basePath: assetsBase,
-  rootDir: path.resolve(import.meta.dirname, '../../../..'),
+  rootDir: path.resolve(import.meta.dirname, '../../..'),
   allow: [
     'demos/data-model-forms/app/assets/**',
     'demos/data-model-forms/app/actions/registration/registration-form.ts',
     'demos/data-model-forms/app/actions/registration/styles.ts',
+    'demos/data-model-forms/app/data/account-schema.ts',
     'demos/data-model-forms/app/routes.ts',
     'packages/*/src/**',
   ],
