+++ demos/data-model-forms/app/data/account.ts
@@ -1,15 +1,7 @@
 import * as s from 'remix/data-schema'
-import { email, max, maxLength, min, minLength, url } from 'remix/data-schema/checks'
 import { column as c, table } from 'remix/data-table'
 
-export const Account = s.object({
-  id: s.string(),
-  displayName: s.string().pipe(minLength(2), maxLength(50)),
-  email: s.string().pipe(email()),
-  age: s.optional(s.number().pipe(min(18), max(120))),
-  website: s.optional(s.string().pipe(url())),
-  password: s.string().pipe(minLength(8), maxLength(72)),
-})
+import { Account } from './account-schema.ts'
 
 const StoredAccount = s.object({
   id: Account.shape.id,
