+++ demos/data-model-forms/app/middleware/render.tsx
@@ -3,7 +3,7 @@ import { createHtmlResponse } from 'remix/response/html'
 import type { RemixNode } from 'remix/ui'
 import { renderToStream } from 'remix/ui/server'
 
-import { assetServer } from '../utils/assets.ts'
+import { assetServer } from '../assets.ts'
 
 export function render() {
   return renderWith(
